feat: Add path update feature. Add replanning feature.
We now consider that robot is moving.
The start position of the path is always updated to the current position of the robot.
It guarantees that the robot will always follow the (newly updated) path.
To update the path start position, we first use path tracing (not replanning).
If path tracing is not possible (i.e. robot is on unserached area),
then wavefront propagation is now done again.